requirements.txt
blinker==1.4
Click==7.0
Flask==1.0.3
Flask-DebugToolbar==0.10.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-slugify==3.0.2
text-unidecode==1.2
Werkzeug==0.15.4
